#202 add api restriction and api doc
