--- src/faucetserver/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-# from .models import Transaction
-
-# admin.site.register(Transaction)
